+++ packages/block-editor/src/components/keyboard-shortcuts/index.js
@@ -115,6 +115,15 @@ function KeyboardShortcutsRegister() {
 		} );
 
 		registerShortcut( {
+			name: 'core/block-editor/stop-editing-as-blocks',
+			category: 'block',
+			description: __( 'Finish editing a design.' ),
+			keyCombination: {
+				character: 'escape',
+			},
+		} );
+
+		registerShortcut( {
 			name: 'core/block-editor/select-all',
 			category: 'selection',
 			description: __(