+++ packages/editor/src/dataviews/fields/content-preview/content-preview-view.tsx
@@ -6,8 +6,6 @@ import {
 	// @ts-ignore
 	BlockPreview,
 	// @ts-ignore
-	privateApis as blockEditorPrivateApis,
-	// @ts-ignore
 } from '@wordpress/block-editor';
 import type { BasePost } from '@wordpress/fields';
 import { useSelect } from '@wordpress/data';
@@ -17,11 +15,10 @@ import { useEntityBlockEditor, store as coreStore } from '@wordpress/core-data';
  * Internal dependencies
  */
 import { EditorProvider } from '../../../components/provider';
+import { useStyle } from '../../../components/global-styles';
 import { unlock } from '../../../lock-unlock';
 // @ts-ignore
 import { store as editorStore } from '../../../store';
-
-const { useGlobalStyle } = unlock( blockEditorPrivateApis );
 
 function PostPreviewContainer( {
 	template,
@@ -30,7 +27,7 @@ function PostPreviewContainer( {
 	template: any;
 	post: any;
 } ) {
-	const [ backgroundColor = 'white' ] = useGlobalStyle( 'color.background' );
+	const [ backgroundColor = 'white' ] = useStyle( 'color.background' );
 	const [ postBlocks ] = useEntityBlockEditor( 'postType', post.type, {
 		id: post.id,
 	} );