+++ packages/format-library/src/math/index.js
@@ -42,16 +42,18 @@ function InlineUI( {
 
 	// Update the math object in real-time as the user types
 	const handleLatexChange = ( newLatex ) => {
-		let mathML;
+		let mathML = '';
 
 		setLatex( newLatex );
 
-		try {
-			mathML = latexToMathML( newLatex, { displayMode: false } );
-			setError( null );
-		} catch ( err ) {
-			setError( err.message );
-			return;
+		if ( newLatex ) {
+			try {
+				mathML = latexToMathML( newLatex, { displayMode: false } );
+				setError( null );
+			} catch ( err ) {
+				setError( err.message );
+				return;
+			}
 		}
 
 		const newReplacements = value.replacements.slice();