+++ test/e2e/specs/editor/various/inserting-blocks.spec.js
@@ -530,7 +530,7 @@ test.describe( 'Inserting blocks (@firefox, @webkit)', () => {
 			.click();
 		await page
 			.getByRole( 'listbox', { name: 'Text' } )
-			.getByRole( 'option', { name: 'Heading' } )
+			.getByRole( 'option', { name: 'Heading', exact: true } )
 			.hover();
 
 		await expect( insertingBlocksUtils.indicator ).toBeVisible();
@@ -571,7 +571,7 @@ test.describe( 'Inserting blocks (@firefox, @webkit)', () => {
 		await page.getByRole( 'button', { name: 'Browse All' } ).click();
 		await page
 			.getByRole( 'listbox', { name: 'Text' } )
-			.getByRole( 'option', { name: 'Paragraph' } )
+			.getByRole( 'option', { name: 'Paragraph', exact: true } )
 			.click();
 		await editor.canvas
 			.getByRole( 'document', { name: 'Empty block' } )
@@ -673,7 +673,7 @@ test.describe( 'Inserting blocks (@firefox, @webkit)', () => {
 			.click();
 		await page
 			.getByRole( 'listbox', { name: 'Text' } )
-			.getByRole( 'option', { name: 'Paragraph' } )
+			.getByRole( 'option', { name: 'Paragraph', exact: true } )
 			.hover();
 
 		await expect(